Extract scan orchestration partial
